// src/types.ts
export type WeatherForecast = {
  weather: WeatherTypes;
  temp: number;
}[];

export type WeatherTypes =
  | 'BLIZZARD'
  | 'CLEAR'
  | 'CLEARING'
  | 'CLOUDS'
  | 'EXTRASUNNY'
  | 'FOGGY'
  | 'HALLOWEEN'
  | 'NEUTRAL'
  | 'OVERCAST'
  | 'RAIN'
  | 'SMOG'
  | 'SNOW'
  | 'SNOWLIGHT'
  | 'THUNDER'
  | 'XMAS'
  | 'UNKNOWN';

/* Created by Whitigol Web Design */